+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/repository/CategoryRepository.java
@@ -1,4 +1,7 @@
 package com.ohgiraffers.adregamdi.place.command.domain.repository;
 
-public interface CategoryRepository {
+import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
