+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/repository/CityRepository.java
@@ -1,4 +1,7 @@
 package com.ohgiraffers.adregamdi.place.command.domain.repository;
 
-public interface CityRepository {
+import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.City;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CityRepository extends JpaRepository<City, Long> {
 }
