+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/repository/DongRepository.java
@@ -1,4 +1,7 @@
 package com.ohgiraffers.adregamdi.place.command.domain.repository;
 
-public interface DongRepository {
+import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.Dong;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DongRepository extends JpaRepository<Dong, Long> {
 }
