+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/repository/PlaceRepository.java
@@ -1,4 +1,7 @@
 package com.ohgiraffers.adregamdi.place.command.domain.repository;
 
-public interface PlaceRepository {
+import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.Place;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PlaceRepository extends JpaRepository<Place, Long> {
 }
