+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/repository/PlaceTagsRepository.java
@@ -1,4 +1,7 @@
 package com.ohgiraffers.adregamdi.place.command.domain.repository;
 
-public interface PlaceTagsRepository {
+import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.PlaceTags;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PlaceTagsRepository extends JpaRepository<PlaceTags, Long> {
 }
