+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/repository/TagRepository.java
@@ -1,4 +1,7 @@
 package com.ohgiraffers.adregamdi.place.command.domain.repository;
 
-public interface TagRepository {
+import com.ohgiraffers.adregamdi.place.command.domain.aggregate.entity.Tag;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TagRepository extends JpaRepository<Tag, Long> {
 }
